api/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.22
pillow==11.3.0
numpy==2.4.2